decawork-it-agent/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
browser-use==0.1.40
langchain-anthropic==0.3.3
langchain==0.3.0
playwright==1.49.0
anthropic==0.41.0
pytest==8.2.0